Move mint transaction modal to separate component
