+++ v2/components/Mint/Mint.stories.tsx
@@ -13,7 +13,6 @@ export const Primary = Template.bind({});
 Primary.args = {
   unstakedSnx: 2000,
   susdBalance: 2000,
-  exchangeRate: 0.25,
   transactionFee: wei(0.01),
   isLoading: false,
 };
